refactor(runner): extract session-identity.ts from session-pool.ts
Move the pure identity, fingerprint, credential-epoch, and pool-key helpers out
of session-pool.ts into a new session-identity.ts, leaving session-pool.ts as the
mutable pool (map, timers, capacity, state, teardown). Pure code move: importers
repointed, no behavior change. Typecheck clean; 1158 runner unit tests pass.
